requirements.txt
dj-database-url==0.5.0
Django
django-allauth==0.42.0
django-braces==1.14.0
django-cors-headers==3.4.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.1
django-environ==0.4.5
django-extensions==2.2.9
django-registration==2.4.1
django-registration-redux==1.6
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
gunicorn==20.0.4
Pillow==7.1.2
pipenv==2018.11.26
postgres==3.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
psycopg2-pool==1.1
six==1.12.0
virtualenv==20.0.21
virtualenv-clone==0.5.4
whitenoise==5.2.
requests
lxml
django-bootstrap4
